[79]-fix-style-and-rename-folder
