+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { CreateSandboxPageComponent } from './pages/create-sandbox-page/create-s
 import { ErrorPageComponent } from './pages/error-page/error-page.component';
 import { HomePageComponent } from './pages/home-page/home-page.component';
 import { CandidateListComponent } from './components/candidate/candidate-list/candidate-list.component';
-import { SandboxListPageComponent } from './pages/sandbox-list-page/candidate-table/sandbox-list-page.component';
+import { SandboxListPageComponent } from './pages/sandbox-list-page/sandbox-list-page/sandbox-list-page.component';
 
 const routes: Routes = [
   { path: '', component: HomePageComponent },
