+++ src/app/pages/sandbox-list-page/sandbox-list-page/sandbox-list-page.component.html
@@ -10,7 +10,6 @@
         <ng-container matColumnDef="select">
           <th mat-header-cell *matHeaderCellDef>
             <mat-checkbox
-              color="primary"
               (change)="$event ? masterToggle() : null"
               [checked]="selection.hasValue() && isAllSelected()"
               [indeterminate]="selection.hasValue() && !isAllSelected()"
@@ -20,7 +19,6 @@
           </th>
           <td mat-cell *matCellDef="let row">
             <mat-checkbox
-              color="primary"
               (click)="$event.stopPropagation()"
               (change)="$event ? selection.toggle(row) : null"
               [checked]="selection.isSelected(row)"
