+++ src/app/pages/sandbox-list-page/sandbox-list-page/sandbox-list-page.component.scss
@@ -1,3 +1,5 @@
+@import '../../../shared/constants/colors.scss';
+
 table {
   width: 100%;
 }
@@ -63,3 +65,6 @@ tr {
   border-bottom: 2px solid #adabad;
 }
 
+::ng-deep .mat-checkbox-checked.mat-accent .mat-checkbox-background,.mat-checkbox-indeterminate.mat-accent .mat-checkbox-background {
+  background-color: $blueColor;
+}
